refactor(turn): split event emitters and resume dispatch for wave 4 (#4113)
Slice A–B of colonizethis_turn wave 4: split turn_resolution_events by
emitter family with stable barrel re-exports, extract diplomacy resume
entry points to turn_resolver_resume.dart, and add repo.turn_lib_file_size
(400 physical-line ceiling) so near-cap modules cannot silently regrow.
