+++ SPEC/program/repo-lint.md
@@ -116,6 +116,7 @@
 | `tool/check_turn_test_resolve_complete.dart` | Forbid nested `requireTurnResolutionComplete(resolveTurnForGame(...))` and direct `resolveTurnForGame(` under `packages/colonizethis_turn/test/**` outside the harness and a short pending-result allowlist; success paths MUST use `resolveTurnComplete` from `test/support/turn_resolver_test_harness.dart` (Refs #4039); rule `repo.turn_test_resolve_complete` |
 | `tool/check_turn_test_fixture_location.dart` | Forbid `*_fixtures.dart` / `*_test_support.dart` under `packages/colonizethis_turn/test/**` outside `test/support/` so fixture modules stay co-located with harnesses (Refs #4039); rule `repo.turn_test_fixture_location` |
 | `tool/check_turn_test_support_loc.dart` | Ratchet ceiling for all Dart under `packages/colonizethis_turn/test/support/` (physical LOC; post-#4039 densify ceiling **5950**, measured ≈5843 with headroom); fail on net regrowth beyond the ceiling (Refs #4039); rule `repo.turn_test_support_loc` |
+| `tool/check_turn_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_turn/lib/**` physical lines per file (wave-4 ≤400; shrink-only grandfather allowlist during transition; generated suffixes excluded); rule `repo.turn_lib_file_size` (Refs #4113) |
 | `tool/check_ai_test_no_duplicate_scaffolding.dart` | Keep `packages/colonizethis_ai/test/**` scaffolding coherent. Three regressions are violations: (1) any `_test.dart` living inside `test/support/` (the scaffolding directory holds helpers only; test files belong in a sibling such as `test/support_test/`); (2) any file named `test_fixtures.dart` (a confusing shadow of the shared `colonizethis_test` `TestFixtures` — AI-specific fixtures must use a disambiguated name such as `ai_planner_fixtures.dart`); and (3) any AI test file that redeclares a top-level `TestFixtures` class instead of importing `package:colonizethis_test/game_test_fixtures.dart`. References to the shared `TestFixtures` symbol are unaffected. Complements the testing-rule shared-fixtures/`test/`-mirrors-`lib/` policy (Refs #3717); rule `repo.ai_test_no_duplicate_scaffolding` |
 | `tool/check_ai_expand_peace_test_shared_fixtures.dart` | Forbid local expand-peace / Phase-8 fixture redeclarations: (1) `expand_phase_planner_*peace*_test.dart` must not redeclare `_ownSnapshot` / `_criticalGame` / `_distractionGame` / `_zeroRegimentGame` / `_peerGame` / `_ownVsPartnerGame`; (2) documented ownSnapshot adopters (`expand_phase_planner_focus_minor_target_*_cases.dart`, `expand_phase_planner_stalled_minor_or_gp_blocker_pivot_test.dart`, `observer_goal_phase_survival_great_power_peace_targets_test.dart`) must not redeclare `_ownSnapshot`; (3) `expand_phase_peace_matrix_sole_gp_blocker_cases.dart` must not redeclare `_ownVsPartnerGame`. Callers MUST import `ownSnapshot` / `buildOwnVsPartnerExpandPeaceGame` (and the other shared Game builders) from `packages/colonizethis_ai/test/support/expand_phase_peace_test_support.dart`. The gate also fails when that shared support file is missing (Refs #3967 / #3997); rule `repo.ai_expand_peace_test_shared_fixtures` |
 | `tool/check_ai_orchestrator_test_shared_fixtures.dart` | Forbid local OW-quota province-list redeclarations in `packages/colonizethis_ai/test/planning/domain_planner_orchestrator_*_test.dart`. A violation is any `const List<String> _gp1OwProvincesBelowQuota` or `const List<String> _gp1OwProvincesAtQuota` declaration inside those pins. Pins MUST import `kGp1OwProvincesBelowQuota` / `kGp1OwProvincesAtQuota` from `packages/colonizethis_ai/test/support/domain_planner_orchestrator_test_support.dart`. The gate also fails when that shared support file is missing. Phase-variant lists (near-quota / develop-only / custom `_gp1OwProvinces`) remain file-local when they are not the canonical below-/at-quota sets. Additionally, pins that call paired Game builders MUST NOT redeclare local snapshot helpers: `buildOrchestratorExpandMinorWarScenarioGame` / `buildOrchestratorPendingCostTradeScenarioGame` → no `_expandSnapshot` (use `buildOrchestratorExpandMinorWarAtWarSnapshot`); `buildOrchestratorGp1TribeNwScenarioGame` → no `_expandSnapshot` / `_colonialSnapshot` (use `buildOrchestratorExpandNwTribeTargetSnapshot` / `buildOrchestratorColonialNwTribeTargetSnapshot`); `buildOrchestratorExpandGpOnlyBlockerScenarioGame` → no `_expandSnapshot` / `_developSnapshot` (use `buildOrchestratorExpandGpOnlyBlockerSnapshot` / `buildOrchestratorDevelopGpOnlyBlockerSnapshot`); `buildOrchestratorExpandAdjacentMinorScenarioGame` → no `_expandSnapshot` / `_developSnapshot` (use `buildOrchestratorExpandAdjacentMinorSnapshot` / `buildOrchestratorDevelopAdjacentMinorSnapshot`); `buildOrchestratorDevelopGpOwnedNwScenarioGame` → no `_developSnapshot` (use `buildOrchestratorDevelopNoColonialTargetsSnapshot`); `buildOrchestratorColonialLiteDeclareWarScenarioGame` → no `_colonialSnapshot` / `_colonialLiteSnapshot` (use `buildOrchestratorColonialNwTribeTargetSnapshot` / `buildOrchestratorExpandNwTribeTargetSnapshot`). Documented diplomatic-scoring COLONIAL tribe adopters (`diplomatic_candidate_scoring_intervention_tribe_tolerance_test.dart`, `diplomatic_candidate_scoring_personality_colonial_divergence_test.dart`) MUST NOT redeclare `_gp1OwProvincesAtQuota` or a brace-bodied `_colonialSnapshot` (use `kGp1OwProvincesAtQuota` / `buildOrchestratorColonialNwTribeTargetSnapshot`) (Refs #3941 / #3997); rule `repo.ai_orchestrator_test_shared_fixtures` |
